Move cache updaters to domain/eventEditor
