+++ assets/prototype/domain/eventEditor/data/entities/index.ts
