+++ assets/prototype/domain/shared/data/queries/index.ts
@@ -1,5 +1,3 @@
-export * from './entities';
-
 export * from './currentUser';
 
 export * from './generalSettings';
